Change InPC of Tamil U and UU from Bottom to Right.
U+0BC1 and U+0BC2 were the only two remaining characters with a
general category of Mc but with a InPC of Top or Bottom. Since they
more frequently than not form ligatures that adds to the width of the
consonant on the right side, making them Right is as good as keeping
them as Bottom, with the added benefit that there would no longer be
any gc=Mc character with InPC of Top or Bottom.