Add Tamil to scx of U+11303 (Grantha visarga).
Apparently a two-dot visarga is attested in Tamil, but the Tamil
visarga, U+0B83 ஃ (aka.  aytham), is quite different, as it's only
historically a visarga.

Examples of the two-dot visarga be seen on page 7 of L2/10-256R
(http://www.unicode.org/L2/L2010/10256r-extended-tamil.pdf) and
pages 2 and 3 of L2/10-407
(http://www.unicode.org/L2/L2010/10407-ext-tamil-follow2.pdf).

Discussions at UTC suggested that the character is unified with the
Grantha visarga.