Clarify the Joining_Type of U+202F NARROW NO-BREAK SPACE.
The character U+202F NNBSP is used in Mongolian, with interesting
joining behavior: it will make the joining character before it final
or isolated, but the joining character after it could become one of
several forms.

Quoting Unicode 7.0 Core Specification, Section 13.4, page 517:

"NNBSP affects the form of the preceding and following letters. The
final letter of the stem or suffix preceding the NNBSP takes the
final positional form, whereas the first letter of the suffix
following NNBSP may take the normal initial form, a variant initial
form, a medial form, or a final form, depending on the particular
suffix."

NNBSP is treated in all major rendering engines to be non-joining,
and the extra magic of the next letter getting shaped correctly for
the context is handled by the font.

This commit does not change anything normative in the UCD since the
character defaulted to Joining_Type=Non_Joining and
Joining_Group=No_Joining_Group anyway, but just clarifies those
properties.