Change joining type of Mongolian baludas to transparent
This is per http://www.unicode.org/L2/L2016/16041r-mongolian.pdf and
a UTC consensus for the change.

We could remove the listing from the file altogether, since the
change in general category will get them to become transparent, but
it seems better to explicitly list them to bring attention to them
having changed.